bugfix fixed a bug that lead to whole site being deleted after reindexing of one page

+++ src/main/java/com/github/dmitriims/posikengine/service/IndexingMonitor.java
@@ -49,7 +49,12 @@ public class IndexingMonitor implements Runnable {
 
                     if (pool.getValue().isQuiescent()) {
                         synchronized (commonContext.getDatabaseService()) {
-                            boolean indexedAnything = commonContext.getDatabaseService().removeDeletedPagesForSite(pool.getKey().getId());
+                            boolean indexedAnything = true;
+                            if(commonContext.isIndexingOnePage()) {
+                                commonContext.setIndexingOnePage(false);
+                            } else {
+                                indexedAnything = commonContext.getDatabaseService().removeDeletedPagesForSite(pool.getKey().getId());
+                            }
                             if(indexedAnything) {
                                 commonContext.getDatabaseService().setSiteStatusToIndexed(pool.getKey().getId());
                             } else {
